Move BoundModule and funcs to a separated module
